[NRRPLT-51470] - New script for the showcase experiments
This script is purposed to help the DevOp run continuously a list of
template experiments on a dedicated server. First, we provide a .json 
which contains all the desired experiments, then we read it, and loop 
through all of them doing the following steps for each:

1. Try to launch the experiment. Keep trying if the server is busy
2. Wait until the experiment finishes or we get a timeout. By default 
the timeout value is set to 1000, as the current 4 experiments are 
timeing out at 840.
3. Repeat until the sun runs out of fuel

Change-Id: I46689d740e71cb6242513000e64ca69848a7303b