[NRRPLT-5091] Add BIBI Support For Custom Timestep
This commit modified the BIBI definition to allow a user to
specifcy a custom simulation timestep to override the default
simulation integration timestep of 20ms.

This value defines how long the brain/robot simulations will be
run before values are collected and TFs are executed. Allowed
values are between 1 and 1000 ms, which is a tentative upper
limit for a semi-responsive frontend  that can be changed in
the future.

The value is optional and will default to 20ms if not present
for backwards compatibility.

Change-Id: Ia4d1ea1e7a40545008c777893b816c99b0bd90ad