[NRRPLT-5284] Add RNG seed parameter to ExDConf.
This patch adds support for specifying an experiment RNG seed, which
will currently be used for Nest and Gazebo.

If the RNG seed is not specified, then a new, random value will be
generated per experiment run. This was the previous behavior due to
an issue with the way the RNG seed was set in the NRP, but the
documentation will be updated to explicitly state this.

Change-Id: Iec82919aa3688b9ffe06d731fd56577a5722af6f