[NRRPLT-5330] Basic XSD syntax/style cleanup.
This patch attempts to cleanup some of the common syntax mistakes
or style errors in the BIBI and ExD schemas. The rationale is to
improve consistency as more developers add (copy/paste) entries
moving forward, as well as to ensure the schema actually meets
expectations (e.g. expected attributes required, etc.).

Most notably:
- default values for elements are not respected if an element is
  optional, remove these to ensure schema meets expectations
- default min and max occurs is 1, remove redundancy/verbosity of
  definitions that imply this is not the case
- default attributes are optional, ensure required values are
  required (default values are valid for optional attributes)
- collapse some complex definitions that are only used in one place
- consistent naming conventions with PascalCase for types, camelCase
  for elements/attributes - remove inconsistent _ cases

Change-Id: If79e133c4b07621071d299080f496e78397ce968