[NRRPLT-5221] Fix screen switching husky initial/reset pose.
The initial position of the husky was set to 0,0,0 which means that
after a reset the Husky appears in the floor until play is pressed
and the physics engine runs a step. This is unique to reset because
of the order of loading during experiment launch vs reset which
would cause the robot to appear properly on launch but not reset.

The position has been set to the same as in all other husky experiments,
meaning it will drop from a small height and everything will work
as expected.

Change-Id: I1d3ca96999b04a54f72c20fdffc56ec4cf2c8dc1