[NRRPLT-5647] Fix CSV robot position TF indexing.
This patch makes the indexing of robot position within the CSV
robot position TF more flexible to support deletions and resets
that can cause the index of the robot to change in the position
message produced by Gazebo.

It attempts to be more efficient by only finding the robot position
if it has changed, rather than each iteration.

Change-Id: I6e4c4a1659f121944c5148eecf101cfa624ee807